requirements/base.txt
pytz==2021.3  # https://github.com/stub42/pytz
python-slugify==5.0.2  # https://github.com/un33k/python-slugify
Pillow==8.4.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.3.0  # https://github.com/evansd/whitenoise
redis==3.5.3  # https://github.com/redis/redis-py
hiredis==2.0.0  # https://github.com/redis/hiredis-py
celery==5.2.1  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.2.1  # https://github.com/celery/django-celery-beat

# Django
# ------------------------------------------------------------------------------
django==3.2.10  # pyup: < 4.0  # https://www.djangoproject.com/
django-environ==0.8.1  # https://github.com/joke2k/django-environ
django-model-utils==4.2.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.46.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.13.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.6  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.1.0  # https://github.com/jazzband/django-redis

# Django REST Framework
djangorestframework==3.12.4  # https://github.com/encode/django-rest-framework
django-cors-headers==3.10.1 # https://github.com/adamchainz/django-cors-headers

anyjson==0.3.3
django-filebrowser-no-grappelli==3.8.0
django-imagekit==4.1.0
django-mathfilters==1.0.0
django-tinymce==3.3.0
fonttools==4.28.2
html5lib==1.1

loguru==0.5.3

pdfkit==1.0.0

pilkit==2.0
python-magic==0.4.24
SQLAlchemy==1.4.27
webencodings==0.5.1
sortedcontainers==2.1.0
prometheus-client==0.12.0
prompt-toolkit==3.0.22
pycparser==2.21
pydyf==0.1.2
pyphen==0.11.0
python-crontab==2.6.0
django-tenant-schemas==1.11.0